Need Help Understanding...
Had to remove "extends TestCase" from this:
public class RoverTest extends TestCase {

I had originally added it to use setUp() method...that would not work
without it.

Later when I wanted to test for exception using:

@Rule
public ExpectedException thrown = ExpectedException.none();
…
@Test
public void throwsExceptionEncouterObstacle() {
	thrown.expect(CannotProceed.class);
	createRoverOnGrid(x, y, direction);
	mars.addObstacleAt(0, 9);
	goBackward();
}

I had to remove it which caused the setUp() not to work.

Refactored to not use setUp() and get exception code to be only failing
test.